Added selected option and uploading multiple files

+++ app/controllers/file_manager_controller.rb
@@ -1,6 +1,12 @@
 # Code for file manager adapted from: adrientoub/file-explorer
 require 'archive'
 require 'pathname'
+
+# 1) Allow for uploading of a folder + select multiple files
+# 2) Select multiple files to delete (tick and checkmark)
+# 3) Allow for folder download
+# 4) Scrolling within the page
+
 class FileManagerController < ApplicationController
   BASE_DIRECTORY = Rails.root.join('courses')
   before_action :set_base_url
@@ -20,12 +26,11 @@ class FileManagerController < ApplicationController
   end
 
   def path
-    absolute_path = sanitize_path(check_path_exist(params[:path].nil? ? "" : params[:path]))
-    instructor_paths = get_all_paths
-    if File.directory?(absolute_path) && absolute_path.in?(instructor_paths)
+    absolute_path = check_path_exist(params[:path].nil? ? "" : params[:path])
+    if File.directory?(absolute_path) && check_instructor(absolute_path)
       populate_directory(absolute_path, "#{params[:path]}/")
       render 'file_manager/index'
-    elsif File.file?(absolute_path) && absolute_path.in?(instructor_paths)
+    elsif File.file?(absolute_path) && check_instructor(absolute_path)
       if (File.size(absolute_path) > 1_000_000 || params[:download])
         send_file absolute_path
       else
@@ -42,53 +47,51 @@ class FileManagerController < ApplicationController
 
   def delete
     absolute_path = check_path_exist(params[:path])
-    if File.directory?(absolute_path)
-      FileUtils.rm_rf(absolute_path)
-    else
-      FileUtils.rm(absolute_path)
+    if check_instructor(absolute_path)
+      if File.directory?(absolute_path)
+        FileUtils.rm_rf(absolute_path)
+      else
+        FileUtils.rm(absolute_path)
+      end
     end
-    new_absolute_path = File.expand_path("..", absolute_path)
-    path_split = new_absolute_path.split("/")
-    courses_index = path_split.index("courses")
-    result = path_split[(courses_index + 1)..(path_split.length - 1)].join("/")
-    populate_directory(new_absolute_path, "")
-    new_path = "#{file_manager_index_path}/#{result}"
-    redirect_to new_path
+    flash[:success] = "File(s) deleted."
   end
 
   def rename
     absolute_path = check_path_exist(params[:relative_path])
-    dir_name = File.dirname(params[:relative_path])
+    if check_instructor(absolute_path)
+      dir_name = File.dirname(params[:relative_path])
 
-    if params[:new_name].empty?
-      raise ArgumentError, "New name not provided,
+      if params[:new_name].empty?
+        raise ArgumentError, "New name not provided,
         new name cannot be blank"
-    end
+      end
 
-    unless params[:new_name].match(/^[a-zA-Z0-9_-]+(\.[a-zA-Z0-9_-]+)?$/)
-      raise ArgumentError, "Invalid characters. Only letters,
+      unless params[:new_name].match(/^[a-zA-Z0-9_-]+(\.[a-zA-Z0-9_-]+)?$/)
+        raise ArgumentError, "Invalid characters. Only letters,
         numbers, underscores, and hyphens are allowed."
+      end
+
+      new_path = safe_expand_path("#{dir_name}/#{params[:new_name]}")
+      parent = new_path.split[0..-2].join('/')
+
+      original_has_extension = !File.extname(absolute_path).empty?
+
+      if original_has_extension && File.extname(params[:new_name]).empty?
+        raise ArgumentError,
+              "You cannot name a file a folder"
+      end
+      if !original_has_extension && !File.extname(params[:new_name]).empty?
+        raise ArgumentError,
+              "You cannot name a folder a file"
+      end
+
+      raise ArgumentError, "A file with that name already exists" if File.exist?(new_path)
+
+      FileUtils.mkdir_p(parent)
+      FileUtils.mv(absolute_path, new_path)
+      flash[:success] = "File successfully renamed"
     end
-
-    new_path = safe_expand_path("#{dir_name}/#{params[:new_name]}")
-    parent = new_path.split[0..-2].join('/')
-
-    original_has_extension = !File.extname(absolute_path).empty?
-
-    if original_has_extension && File.extname(params[:new_name]).empty?
-      raise ArgumentError,
-            "You cannot name a file a folder"
-    end
-    if !original_has_extension && !File.extname(params[:new_name]).empty?
-      raise ArgumentError,
-            "You cannot name a folder a file"
-    end
-
-    raise ArgumentError, "A file with that name already exists" if File.exist?(new_path)
-
-    FileUtils.mkdir_p(parent)
-    FileUtils.mv(absolute_path, new_path)
-    flash[:success] = "File successfully renamed"
   rescue ArgumentError => e
     flash[:error] = e.message
   end
@@ -102,8 +105,7 @@ private
   end
 
   def populate_directory(current_directory, current_url)
-    directory = Dir.entries(sanitize_path(current_directory))
-    instructor_paths = get_all_paths
+    directory = Dir.entries(current_directory)
     @directory = directory.map do |file|
       abs_path_str = "#{current_directory}/#{file}"
       stat = File.stat(abs_path_str)
@@ -112,12 +114,12 @@ private
         inst = true
       else
         abs_path = Pathname.new(abs_path_str)
-        inst = abs_path.in?(instructor_paths)
+        inst = check_instructor(abs_path)
       end
       {
         size: (if is_file
                  begin
-                   number_to_human_size stat.size
+                   stat.size
                  rescue StandardError
                    '-'
                  end
@@ -144,22 +146,30 @@ private
     unless (current_directory == tested_path) || Archive.in_dir?(tested_path, current_directory)
       raise ArgumentError, 'Should not be parent of root'
     end
-
     tested_path
   end
 
   def upload_file(path)
     absolute_path = check_path_exist(path)
-    raise ActionController::ForbiddenError unless File.directory?(absolute_path)
+    if absolute_path === BASE_DIRECTORY
+      flash[:error] = "You cannot upload files in the root course directory " \
+         "#{view_context.link_to 'here', new_course_url, method: 'get'}" \
+         " if you want to create a new course."
+      flash[:html_safe] = true
+    else
+      raise ActionController::ForbiddenError unless File.directory?(absolute_path)
 
-    input_file = params[:file]
-    return unless input_file
+      if check_instructor(absolute_path)
+        input_file = params[:file]
+        return unless input_file
 
-    all_filenames = Dir.entries(absolute_path)
-    @duplicated_file = all_filenames.include?(input_file.original_filename)
+        all_filenames = Dir.entries(absolute_path)
+        @duplicated_file = all_filenames.include?(input_file.original_filename)
 
-    File.open(Rails.root.join(absolute_path, input_file.original_filename), 'wb') do |file|
-      file.write(input_file.read)
+        File.open(Rails.root.join(absolute_path, input_file.original_filename), 'wb') do |file|
+          file.write(input_file.read)
+        end
+      end
     end
   end
 
@@ -175,38 +185,19 @@ private
     @base_url = '/file_manager'
   end
 
-  def get_all_paths
+  def check_instructor(path)
     current_user_id = current_user.id
+    path = Pathname.new(path)
     cuds = CourseUserDatum.where(user_id: current_user_id, instructor: true)
     courses = cuds.map do |cud|
       Course.find_by(id: cud.course_id)
     end
-    paths = courses.map do |course|
-      Pathname.new("#{BASE_DIRECTORY}/#{course.name}")
+    courses.map do |course|
+      course_path = Pathname.new("#{BASE_DIRECTORY}/#{course.name}")
+      if path == course_path || Archive.in_dir?(path, course_path)
+        return true
+      end
     end
-    children = paths.map do |path|
-      get_all_children(path)
-    end
-    (children + paths).flatten.compact.uniq
-  end
-
-  def get_all_children(path)
-    return [] unless path.exist?
-    return [path] unless path.directory?
-
-    children = path.children
-    all_children = [children]
-    children.map do |child|
-      all_children += get_all_children(child)
-    end
-    all_children.flatten
-  end
-
-  def sanitize_path(path)
-    allowed_pattern = %r{\A[a-zA-Z0-9_\-/]+(\.[a-zA-Z]+)?\z}
-    unless path.to_s.match?(allowed_pattern)
-      raise ActionController::ForbiddenError(path, 'is not in a valid format')
-    end
-    path
+    false
   end
 end
